feat: refactor inputs, handle nested tuple type
