+++ packages/components/src/components/queryBuilder/objectInputs.tsx
@@ -1,6 +1,6 @@
 import { QueryBuilderInput } from ".";
 import { Property } from "../../utils/playground-request";
-import { TupleItem } from "./arrayInputs";
+import { InputItem } from "./InputItem"
 
 interface ObjectInputsProps {
   props: Property[]
@@ -12,22 +12,20 @@ interface ObjectInputsProps {
 
 export const ShowNoInputTypes = ["null", "undefined"]
 
-export const ObjectInputs: React.FC<ObjectInputsProps> = ({ props, setInputType, setInputValue, getInputFromState, indent }) => {
-  return (
-    <div className="border-secondary" style={{ marginLeft: indent ? "0.75rem" : "0", borderLeftWidth: indent ? "2px" : 0 }}>
-      {props.map((prop, idx) => (
-        <TupleItem
-          key={idx}
-          types={prop.type}
-          indent={indent}
-          input={getInputFromState(prop.name)}
-          inputType={getInputFromState(prop.name)?.type}
-          name={prop.name}
-          prop={prop}
-          setInputType={(type) => setInputType(prop.name, type)}
-          setInputValue={(value) => setInputValue(prop.name, value)}
-        />
-      ))}
-    </div>
-  )
-}
+export const ObjectInputs: React.FC<ObjectInputsProps> = ({ props, setInputType, setInputValue, getInputFromState, indent }) => (
+  <div className="border-secondary" style={{ marginLeft: indent ? "0.75rem" : "0", borderLeftWidth: indent ? "2px" : 0 }}>
+    {props.map((prop, idx) => (
+      <InputItem
+        key={idx}
+        types={prop.type}
+        indent={indent}
+        input={getInputFromState(prop.name)}
+        inputType={getInputFromState(prop.name)?.type}
+        name={prop.name}
+        prop={prop}
+        setInputType={(type) => setInputType(prop.name, type)}
+        setInputValue={(value) => setInputValue(prop.name, value)}
+      />
+    ))}
+  </div>
+)
